refactor(review): move review to entities
