+++ apps/frontend/src/entities/review/ui/ReviewCard.vue
